Refactor each class/struct/interface into its own file in an appropriate folder
