+++ ChessAI/DataClasses/IDirectionBoard.cs
@@ -1,0 +1,9 @@
+namespace ChessAI.DataClasses
+{
+    public interface IDirectionBoard{
+        byte GetFieldsToEdge(int position , int direction);
+        bool IsFieldOccupied(byte position);
+        bool IsFieldEnemouriouslyConqueredByEvil(byte position);
+
+    }
+}
