+++ ChessAI/DataClasses/Move.cs
@@ -1,0 +1,8 @@
+namespace ChessAI.DataClasses
+{
+    public readonly struct Move{
+        public readonly byte StartPos;
+        public readonly byte EndPos;
+        public Move(byte startPos , byte endPos){ StartPos = startPos; EndPos = endPos; }
+    }
+}
