+++ ChessAI/DataClasses/Piece.cs
@@ -1,0 +1,7 @@
+﻿namespace ChessAI.DataClasses {
+    // inspiration Source https://www.youtube.com/watch?v=U4ogK0MIzqk
+    public class Piece {
+
+
+    }
+}
