--- ChessAI/ExampleClass.cs
@@ -1,8 +1,0 @@
-namespace ChessAI
-{
-    public class ExampleClass
-    {
-        public static int Add(int a, int b) => a + b;
-        public static int Minus(int a, int b) => a - b;
-    }
-}
