+++ ChessAI/GameController.cs
@@ -1,0 +1,9 @@
+using ChessAI.DataClasses;
+using  ChessAI.IO;
+namespace ChessAI
+{
+    public class GameController
+    {
+        
+    }
+}
