+++ ChessAI/IO/IO.cs
@@ -1,0 +1,7 @@
+namespace ChessAI.IO
+{
+    public class IO
+    {
+        
+    }
+}
